refactor: 클래스 예약 리듀서 하나로 통합 및 #37
