--- src/types/class.ts
@@ -1,8 +1,0 @@
-export type Class = {
-  class_id: string;
-  category: string;
-  title: string;
-  location: string;
-  price: number;
-  image: string;
-};
