--- src/types/date.ts
@@ -1,3 +1,0 @@
-type DateTypePiece = Date | null;
-
-type DateType = DateTypePiece | [DateTypePiece, DateTypePiece];
